prometheus-mysql-exporter: Abort early and capture db integrity issue
Sometimes, when zarcillo has inconsistent data (a section has been
added that doesn't exist, or there is a typo, or some other issue)
NULL values are returned.

Rather than trying to work with invalid data, capture the exception
and return a nice error to the logs.

Also, remove old noqa that was added just for CI python2
compatibility.

Change-Id: Ifef78ee47a716c940e9437c3a319f863e989b1b4