hieradata/common/service: fix syntax error in Icinga config
The check command and parameter need to be separated by !
not : in this context. An easy mistake since it's a port number
that is usually always separated by : in other contexts.

follow-up 473845b2792

Currently this causes an Error in Icinga config so it was alerting.

Error: Service check command 'check_https_on_port:8090"..
 not defined anywhere!

from: icinga -v /etc/icinga/icinga.cfg

which meant other monitoring changes couldn't be applied.


Change-Id: Icbb9f5a735b81a2edb37e521492c557a8cef66f5