docker::baseimages: improvements to build-bare-slim
We are now in the situation where we won't need a second layer for these
images that is wikimedia-specific. So add the LC_ALL setting which was the
last missing piece from the openbootstrap-vz images.

Also improve output adding a note about failing apt updates during
debuerreotype-init that can be safely ignored.

Bug: T281596
Change-Id: If59b604ad00c0758c4e085c92c90604ad118fc21