docker-registry: Add caching config for nginx
Allow to cache immutable blobs and manifests on disk to prevent them
from being pulled from swift over and over again.

I needed to reorder the named location redirect to be able to add a
regex sub location (not possible inside a named location).

Cache details are hardcoded currently as this is, in first place, just
for testing the approach. It currently tries to cache blobs/manifests
for 7 days with a limit of 10GB on disk. Byte range caching is enabled
and parallel equal reqests are held back for 60s (waiting for the first
one to finish and serve from cache then).

Bug: T264209
Change-Id: I0dd55940cc446b9771264dca8296691258c844c0