P:pki::multirootca: drop nrpe check for expired certs
It is not really viable to check expired certificates from the database
directly as there is no way to easily determine if a specific CN belongs
to a decommissioned or broken node

Most https checks already check for certificate expiry so we should
detect errors early.  Further a failure to renew a certificate will cause
a puppet failure.

Hosts: pki1001.eqiad.wmnet
Change-Id: I34edb868e353ecad6ff1ab3592efde60ff642acf