profile::reportupdater::jobs: move logs to /tmp/reportupdater-logs
The bigtop::hadoop::directory class seems to create the parent path/dirs
without chowning it as well. Since changing its behavior may be a big
change for a lot of other profiles (like the master ones) I am inclined
to test a simplified /tmp/reportupdater-logs directory and see how it
goes. If needed we'll be able to move to something different later on.

Change-Id: Ie38994b2ef6a647422d6b72b35d1229a0684b0e5