prometheus: Add dependency between varnish exporter and varnish service
The aim is to ensure that the varnish exporter starts and stops when
varnish itself does so. The options available to enforce the
start stop order and behavior are Wants, Requires or BindsTo.

While the first two are good options, they lack the preferred
enforcement i.e they do not ensure that a unit may never be in
an active state without the specified dependent unit being in an
active state. For that reason BindsTo is a better option with a
combination of the option After which will ensure the varnish exporter
only starts after varnish has successfully started.

Including varnish service as a value to WantedBy option in the Install
section will also ensure that the varnish exporter is started whenever
Varnish service is started.

Hosts: cp3050.esams.wmnet,cp3051.esams.wmnet
Bug: T283660
Change-Id: I56a4b56e3387d930d2dee69afa69fb9151b3e5f2