dbbackups: Move s2 from db2098 to db2097, reimage db2098 to buster
We are going to move current stretch backup source from s2, which
us currently on it own, togheter with s1 on db2097 (also stretch).

This will free up db2098, so prepare to reimage it to buster so
it can contain a mariadb 10.4 backup source (passive for now).

Update dependencies so we continue doing stretch backups for now.

Disable notifications on both hosts so they don't complain until
they are fully restored.

Bug: T280979
Change-Id: I140cc21b702f54eaf440933e4ea37ba1e44ecf11