docker_registry_ha: Enable local nginx cache by default
Caching of immutable objects (blobs and manifests by digest) showed good
results in terms of that it lowers traffic towards swift and increases
the available bandwith to serve clients. Enable it by default for all
docker-registry nodes.

The cache is pretty small currently (10GB) and is kept for 7 days max.
These are just guesses based on available disk, so we might be able to
further extend the caching period.

Bug: T256762
Hosts: O:docker_registry_ha::registry
Change-Id: I60be905ece6500dccd12509e72cbf7edb89df657