airflow test - ensure analytics instance is absent, add analytics-test
This will test the ensure => absent capabilites of the airflow
module, and will recreate the airflow instance on an-test-coord
as analytics-test, to avoid confusion with the analytics one
we will eventually create in analytics prod.

Also create dags symlink if dags_folder is not in airflow_home,
else manage it as a dir.

Bug: T272973
Change-Id: I635b13b0d452f2bab18ab4aff2ba37c681f94229