mariadb::instance - allow passing extra configs from hiera
This allows non-parameterized mysqld config values to be passed in and rendered
using default $template mariadb/instance.cnf.erb.  This is now used by
profile::mariadb::misc::analytics::multiinstance to set per-instance
mysqld configs via hiera without having to write new puppet code or
new ERb templates.

Apply explicit_defaults_for_timestamp: on for the analytics_meta backup
instance on db1108.  This is already on for the analytics_meta
non multi instance mysqld on an-coord100[12], and it should
be on for this backup replica too.

Bug: T272973
Change-Id: I0b0cc9598801a51241fe5a0bc31b59ca449957d1