Set up airflow-analytics on an-launcher1002
db_user and db_password have already been added in private repo.

Before this patch is merged, we should create the airflow_analytics database
on an-coord1001, an-coord1002 and db1108 and add GRANTs accordingly.
Then we can run airflow-analytics db upgrade.

Also fix a couple of typos in clean_logs script.

Bug: T272973
Depends-On: Icf504b3ce91f9466c6a1e9d5ab65de11b85e3828
Change-Id: I69097c4e3a3dc54789d0d6a62785537399a937be