sudo: drop keep_env option
STY and TMUX variables where added to this list recently however we have
now fixed the issues that required this so they can be dropped.

HOME was added back in 2014 when debian changed there defaults. However
we have had reports of this causing issues (see task) and it seems best
practice to stick with the debian defaults unless there are pressing
reasons.

There is no work around once this CR goes in place so users
relying on this functionality will need to consider new work flows.

Bug: T275852
Change-Id: I673c6c1cf401b2931cbd94392e97bf36560be21a