P:idp::client::http:site: add support for same site cookie
A version of libapache2-mod-auth-cas has been built and deployed to
support both CASCookieSameSite and CASCookieSecure.  This CR adds
support to the puppet modules so we can configure theses values.

We default to `CASCookieSameSite Lax` as this seems the most sane value
and is already the default in common browsers

[1]https://hacks.mozilla.org/2020/08/changes-to-samesite-cookie-behavior/
[2]https://web.dev/samesite-cookies-explained/
   #explicitly-state-cookie-usage-with-the-samesite-attribute

Bug: T264605
Hosts: cumin:R:profile::idp::client::httpd::site
Change-Id: I53fcf93c42d4a587fd0bedc035012db609e86cfa