reimage: raid0.default_layout=2 for all installers
raid0 reimages were failing for hosts with drives of different sizes.
Ultimately this is because the different drive sizes led to their being
more than one "zone", and the linux kernel lacks the requisite
information to automatically know which layout is correct.

See https://github.com/torvalds/linux/commit/c84a1372df929033cb1a0441fb57bd3932f39ac9

The linux kernel does not have sufficient information to decide which
layout to use; basically, any chunk of data written by a post-3.14
kernel ends up assigned to a different device than in pre-3.14 kernels.
This means that which layout is "correct" depends on the version of the
kernel that wrote the original data. Basically, we'd want layout 2 for
anything post-3.14 and layout 1 if re-imaging a pre-3.14 server in a way
that doesn't wipe every partition (i.e. we still have some data we care
about).

See https://phabricator.wikimedia.org/T280382#7129110 for where we
decided to make raid0.default_layout=2 the default for all installers

Note the associated ticket is WDQS-specific, because we first ran into
this issue in the course of working that ticket, but this change affects
reimages for all services, not just WDQS.

Bug: T274788
Change-Id: I8422852c08a225fa4a2e3e4014171471ed8d7d29