gitlab: add Bacula backup::host class to role
There are 2 things needed to have Bacula backups on a host.

First, include the backup::host class, what we do here,
to add a Bacula agent and needed config.

Second, adding a backup::set which is a collection
of file pathes to be backed up. It can be an existing one
for standard locations or a new custom one specific to this
role.

A custom one needs to be added in the Bacula server role, then
it can be used here.

Bug: T274463
Change-Id: I94fd2d98971c4ffb27b181bd0262737fe512a101