bacula/gitlab: add a backup::set for gitlab and use it
- define a new backup::set on the Bacula server for gitlab,
  consisting of /srv/gitlab-backup/latest (per T274463#7130085)
  and /etc/gitlab/config/latest_backup (per T274463#7135022).

- use the new set in the gitlab profile, follow-up after I94fd2d98971c4f
  to actually activate backups (and monitoring for it), with the
  newly available "Daily" full backups policy

We have to keep an eye on Icinga alerts when adding this the first time.
It might alert because it checks if backups have happened succesfully
but does so already before the first run has been completed.

But in that case we can add the host on a temp. ignore list to silence
them, no big deal either way.

Bug: T274463
Change-Id: I63bfaa59b71b8ab97f1a428a2776ae3d5bdab99c