galera: clear up confusing xtrabackup parameter
Since this is actually just running a mysqldump script and then shipping
the files to the bacula director, get rid of the confusing xtrabackup
parameter unless we actually start using it with the bpipe method.

We have so far decided the databases are likely not big enough to worry
about improving performance by using mariabackup (despite now having
the executable). We can perhaps revisit later.

Bug: T284157
Change-Id: Ied56baa3d4262dd7abc66e58dcffafde036985ac