Deploy the host keytab directly in the profile::base::cuminunpriv profile
profile::kerberos::keytabs is a general purpose mechanism which allows
to deploy a variety of keytabs which are configured via Hiera. But for
rootless Cumin there's always the same keytab, so directly specify so
that we don't need to configure it in Hiera for every role that gets
managed via rootless Cumin.

Hosts: sretest1001.eqiad.wmnet
Bug: T244840
Change-Id: I077cc7693136d8fcd8483886a80e2940091dcc7d