deployment::rsync:: temp allow dumping files from miscweb1002
Temporarily adding an rsync server module that allows
connecting and writing data from miscweb1002.

This is to copy the data behind https://static-bugzilla.wikimedia.org/

Later on it will be moved to dumps.wikimedia.org.

Currently there is not enough space there or on miscweb* itself to make
a second local copy of it. miscweb* ganeti VMs are running low
just like the dumps servers.

Bug: T281538
Change-Id: Ib11bbb384112ded5052eab94ade58ad98be89ef9