puppetmaster: update the private repo pre-commit hook to check staged
Currently the pre-commit hook checks the state of the entire hiera
directory as it exists on disk.  This can cause issues where the
following happens:

    * Change a file locally with an issue
    * Staged for commit
    * Commit failed with yamllint complaining
    * Fix the file on disk
    * Forgot to stage the new change
    * commit again
At this point the CI check checks the fixed version on disk and passes.
However the staged version and the one that actually gets commited is still
the version with the error.

This CR updates the check so that it checks the file as it exists in
its staged form.  Further we also update the logic to only check files
which have changed instead of all yaml files

Finally we add an additional check to ensure all files in hieradata
have a .yaml extension

Hosts: C:puppetmaster::gitclone
Bug: T278187
Change-Id: If0d1972e5fd88e66248822cc796e2d0f39acf914