beta: add warning motd and link to term of uses
It seems we should highlight that access to the beta cluster instances
are subject to the WMCS term of uses. Some people might not be aware of
it, specially when they are not used to WMCS and solely access beta.

While at it add a warning about avoiding unpuppetized changes.

The profile::beta::motd class is applied to all instances of the
deployment-prep via Hiera `classes`.  If I get it rights, values are
merged and instances or hiera configuration via Horizon already using
`classes` would simply be merged in.

Bug: T100837
Change-Id: I81c4429b2335627bfac685a4bcb4dd527f87fe84