Allow mgmt range to connect to tftp servers.
This is to support installing 40 servers today in eqiad, all of
which need firmware upgrades via FTP.  A temporary rule was added
to the Juniper SRX firewall to allow this traffic from mgmt to
produciton earlier on, however it turns out the iptables rules on
the instal1003 machine itself won't allow TFTP connections from
the mgmt range being used (10.65.1.0/24).

This should add a tempoary allow for that range to connect.  The
existing rule matches $PRODUCTION_NETWORKS, but it did not seem
appropriate to augment that with this mgmt range.  As such a
separate rule was added.

Like the SRX firewall change this will be undone tomorrow, and some
time/effort will be required to set up a long-term way to support
this connectivity.

Bug: T283771
Change-Id: I4b03ef87705ea36e8fd0625092f4efd49797f205