nfs prometheus: change to strings for dir sizes
We changed from ints to floats at some point. I don't think either is
right. The value is returned as intended from du, and it is preserved
when it is a string. Additionally, add some shell options for safety and
remove errant filters, and fix the timeout. 10m just caused data loss.

Bug: T284964
Change-Id: Ia3130f6dd870cd978397e57faafa3393bcff326f