mediawiki::web::yaml_defs: fix virtualhost port.
On physical servers, we run apache as root, then drop privileges to
www-data while running. So we can bind on the privileged port 80.

In kubernetes, we don't want that, so we run apache as an unprivileged
user and bind to a non-priviledged port picked by the env variable
APACHE_RUN_PORT. So we modify the content of the virtualhosts for the
sake of running them on kubernetes.

Please note this approach was preferred to modifying the apache configs
to use $APACHE_RUN_PORT directly as that would have implied deeper changes
to the in-production apache configuration.

Change-Id: Icbaa9a520b693e971cb44011279a600225423c04