postgres::slave: remove problematic auto-replicate
For databases of any notable size, this pg_basebackup command will
fail due to puppet timeout. What's more, because of how the
pg_basebackup command runs two processes in parallel (one for import
and one for syncing with the updates since the start of the backup
process), the import process will be killed after timeout but the
update process will continue to run which leaves things in an
indeterminate state.

This failure means that reimaged postgres replicas will always fail as
far as wmf-auto-reimage is concerned.

Rather than try to manage a 2h+ process via puppet, remove this step
entirely and have the updates be part of a manual process of
provisioning new replicas using a cookbook.

Bug: T285476
Change-Id: I45232e1ef81dc3229e92beed9fa2ceda2d563014