Add helper tool for returning a user's current TGT
This traverses cas_audit log file(s) for the TGT of the current user
(only the oldest entry is kept). The cas audit log anonymises the TGT
name to some extent (like TGT-5-*****03sNubRrSI-idp-test1001), but
we can find out the full name by querying memcached itself.

This tool is a workaround: The ssoSessions endpoint would allow us to
query a user's TGT, but the underlying session update isn't implemented
for the memcached backend.

This will be needed for the logout cookbook (so that we can terminate a
user's session via the ssoSessions actuator)

Bug: T283242
Change-Id: Ia90dfbb466d60ec3392a1411f5cdb79cc52380db