profile::gitlab: ensure backup dirs exist in production
Like I98677b8f89a2a and If15800f97d11dc04 but done in the
profile used in production.

Just ensure the backup dirs exist in prod so that we can move
forward with Bacula backups.

Assuming at some point we might merge and use the module in both
prod and cloud.

Bug: T274463
Change-Id: I96c86e444a2b4e8dfcef3d439001f50ef49d9c99