mariadb: Monitor pt-heartbeat for expected status.
Ensure that the service is in the expected state.
To begin with, the alert is not critical, and will notify in the
we can upgrade it to a first-class alert.

Bug: T285079
Hosts: db1157.eqiad.wmnet,db1175.eqiad.wmnet
Hosts: db2105.codfw.wmnet,db2127.codfw.wmnet
Hosts: pc1007.eqiad.wmnet,pc2007.codfw.wmnet
Change-Id: Ia8a46badcf40f2b086060961616eebf501e51893