swift: Only run swiftrepl-mw in the active datacenter
Use conftool to check which datacenter is the active one in etcd and
then skip running if the script is in the wrong one.

This will allow us to enable the timer in all datacenters and have it
automatically switch to the active one post-switchover.

I tested this by running the newly added bash lines on
ms-fe1005.eqiad.wmnet, where it printed nothing, and then on
ms-fe2005.codfw.wmnet, where it said to run the script in eqiad instead.

Bug: T285373
Change-Id: I2f2e58258b3b3c92dc1cfb76f5c9cefff8d43904