spec_helper: switch to rspec mocha and add rspec_parrallel arguments
The combination of puppet autoloading modules and the layering of monkey
patches in rpsec-puppet, means that some of the more complex test around
testing ruby code have issues.  When testing from the root of the puppet
repo instead of changing into module directory to run the test rspec is
unable to find custom providers and facts

As such we add the affected ruby paths directly to the LOAD_PATH rspec
uses.  We also ensure we add the root directory before the custom
libraries to ensure what puppet needs for autoloading has precedence in
the LOAD_PATH

This can be tested locally with:
   bundle exec rake global:parallel_spec

Change-Id: I6a0a0e16972ff99579b8634db8d28eb051e51a84