Bump envoy timeout for restbase
Similar to If487a1b81bb37d963d6620dc9432752dae46b6d7

This matters for parse request that passthrough to Parsoid, which could
take longer than 10s.  Requests for cached parses and serialization are
fast enough that it doesn't seem to matter for them.

This is an alternative to Iffec5fee0d157dee08dec0886c5badc7913b5dc5
which proposes using restbase-for-services for VE's VirtualRestClient
calls, which has a timeout of 122s.

Bug: T279825
Change-Id: Ib8f7e32d22a7089c12974fc06619330c986a7e2b