varnish: add counters for Varnish SLI
Add two new Prometheus counters, varnish_sli_all and varnish_sli_good.
The former counts all requests, while the latter keeps track of the
requests that spent less than 0.1s within Varnish and did not produce an
error.

The 0.1s threshold, as well as the purposedly vague definition of
"producing an error" (Error or FetchError VSL tags present) are details
that may need to be revisited in the future.

Bug: T284576
Hosts: cp3050.esams.wmnet
Change-Id: I68b97d8e7572b4d5723782f3c01f8892ccf0b4b7