cinderutils.ensure: take into account filesystem size
We were not taking into account the space used by the filesystem, that
reduces the available space of the partition. So a 40GB cinder volume
would end up not matching an ensure of 40GB because there's 10MB used
for the filesystem itself.

This hardcodes the 10MB (works for our current sizes, might need
adapting in the future).

Change-Id: I23ae893ea3d6c9ade55b324bc083746b07093800
Signed-off-by: David Caro <dcaro@wikimedia.org>