Don't show Kerberos ticket info in general
This was initially introduced when Kerberos was only used on the analytics
hosts, but these days we also have servers which are mostly part of the
Kerberos realm for remotely executed commands. Printing the banner on
those is confusing for users who login for something unrelated (e.g.
on the apt* servers), so default to not printing the banner (since we'll
be using Kerberos more and more) and enabling it explicitly for select
analytics use cases (hadoop clients, stat* hosts, Airflow) and the
cuminunpriv* hosts.

Hosts: C:profile::kerberos::client
Bug: T244840
Change-Id: I9a2e9ff22a0f470525df4dd2bc0b84952d893b9d