dbbackups: Temporarily disable s4 snapshots to prevent conflict with dumps
We have altered slightly the schedule of snapshots and dumps to fit
better the scheduled datacenter switchover. Because s4 dumps can take a
long time due to T275268, delay its execution until a manual run
is done tomorrow UTC morning.

This is a temporary change and should be reverted after 19 UTC.

Bug: T284897
Change-Id: I6a890d71409a997199e68025835cded77f5e6926