gitlab: move service IP settings from common to DC level
The service IP used by a gitlab machine can't be common
to all machines using the gitlab role once we start adding
it to a second machine in another datacenter as we are about
to do with gitlab2001 in codfw.

So these part of the settings need to move from "common" Hiera
to eqiad Hiera and it should be a complete noop for existing gitlab1001.

Though without this if we'd apply the gitlab role on gitlab2001
we would get an IP conflict.

Bug: T285456
Change-Id: I7a5db2f952c62fe7b679ae9ca3f65912c680048c