gitlab: add parameter for active_host, limit backups to it
We want to add a second host using the gitlab role but we do not
want to duplicate the size of our backups.

We just want to backup the host that is currently the active host
and we want to be able to switch the active host by editing a single
value in Hiera.

So we are adding a parameter $active_host and compare it to FQDN
and only add backup::set if they are equal.

Bug: T285456
Change-Id: I696a9dee172edeea65ba1047465b68a9e43b09af