elastic: Switch to nginx-light
Nothing in /etc/nginx/sites-enabled/production-search-eqiad* and tlsproxy
itself seems to need any of the modules from the -full flavour (which gets
picked by default). So switch to nginx-light to reduce the attack surface
and libraries loaded.

relforge and cloudelastic are already switched and working fine with -light.

Hosts: elastic1045.eqiad.wmnet
Bug: T164456
Change-Id: I96c6a174404f7c14700fbcbe6de13a768fb3c78d