ml_k8s::master: add profile::kubernetes::node
The ML team would need to run some containers on the k8s master nodes,
for example bird to support calico.

This is a prototype to see if kubelets can run on master nodes. Before
merging, the target nodes/vms will need a new /dev/vdb virtual disk
(10g should be enough) for docker and also all the tokens in the
puppet private repo.

Bug: T285927
Change-Id: If6ae818a40045a6390dff07e1212561a4f0fd573