ceph.keyring: make sure the bootstrap dir exists
As we can't declare the keyring dir dynamically in the resource,
because that will create dependency cycles (ex. with /etc/ceph), we
instead explicitly create the bootstrap dir there.

This is a bit flakier in the sense that if we change the path in
hiera it will fail to run, but I don't see any easy way out without
refactoring the modules.

Bug: T285858
Change-Id: I408e5b4ce275bb24a9bcc6c6473d053a499dea3f
Signed-off-by: David Caro <dcaro@wikimedia.org>