kubernetes: centralize the creation of /etc/kubernetes
This is a step to allow profile::kubernetes::node to be deployed
on master nodes (to have a kubelet running on them).
The current settings for /etc/kubernets are more restrictive on
master nodes, but it looks like subdirs have proper permissions
so it should be fine to have root:root 755 across all nodes.

Bug: T285927
Change-Id: Iaaf05697740fdad1ceb8ad8f2e6a56eab7a17c5c