mailman3: Discard all mails with a X-Spam-Score >= 6
This rule was enabled in Mailman2 but was accidentally missed in the
Mailman3 migration.

All mail currently goes through SpamAssassin, which sets a X-Spam-Score
header with the number of `+` characters indicating the integer score.
If there are 6 or more `+`, then Mailman3 will automatically discard
the mail. Individual lists can increase this threshold by setting a
header check for a lower spam score (e.g. 4 plusses), but cannot raise it.

Bug: T286218
Change-Id: I58ffd367ffdbac493075e3e007029cb53abff862