P:configmaster: update disc state to match post dc switch over state
This CR updates the expected DNS discovery state to match the state as
it exists post-DC switch over.  i.e. the primary site is now codfw and
not eqiad.

One thing i did notice while drafting this change is that it seems some
services which would normally be multi site are currently only running
in codfw. Need someone from serviceOps to validate this is intended (see
comments for examples)

Hosts: puppetmaster1001.eqiad.wmnet
Bug: T286231
Change-Id: I40c8ab72b2ef74b997eab50ca49095f1cb1c3922