metricsinfra: Add HAProxy for distributing http traffic
This is in preparation of splitting alert manager to live in a separate
VM, as it adds the ability to route prometheus.wmflabs.org/cloud/
(prometheus itself) and prometheus.wmflabs.org/.alertmanager/ to
separate backend machines, and adds the possiblity to have multiple
hosts for one service type. This is designed to be proxied via the
shared Cloud VPS proxy, which deals with TLS termination for
metricsinfras behalf.

Bug: T286335
Change-Id: I12de7d03f135cb91d13e152679315d2da3f62b54