tcpircbot: Reduce max length of an IRC message.
In T285709 we found that when a message is split into multiple lines,
some characters are lost in between, meaning that each line we're
sending is too long, and is getting truncated to fit the IRC 512-byte
cap for the whole line (likely when the server prefixes our PRIVMSG
command with our hostmask). Lowering BUFSIZE means we'll break the long
TCP message into shorter lines, avoiding lossage in that truncation.

It doesn't fully address T285709 though, because it's still the case
that when tcpircbot is run as logmsgbot, only the first line is prefixed
with !log and thus picked up by stashbot to be posted to the SAL. Thus
shrinking BUFSIZE is a partial solution but we'll have to address that
issue separately.

Bug: T285709
Change-Id: I75bf3f480fa2a1b999fa683a5c4cefa6b2ea8824