role::ml_k8s::master: avoid exposing puppet certs
The puppet certs are exposed in all our k8s masters without a real need,
since the use case seems to be related to toolforge.
Since the ML team needs to deploy kubelets on master nodes, this change
forces puppet to skip the exposure of the host's cert/key in the master role,
to avoid a duplicate declaration (and permission settings clash) with the same
on the node's profile (that will be deployed on masters soon).

Bug: T285927
Change-Id: I8d545c0cb2037aa57c828939a07e79390aedd208