Revert "Revert "mediabackup: Install minio on the storage hosts and open port 9000""
This reverts commit 5dae6f5859638e1f7a2815fdf50caed612ebcdb4.

Reason for revert: Redeploy after fixes

Storage hosts will run mainly minio. Install minio package and
make sure service is running - which requires it being configured
and TLS certificated generated before start.

In order for the service to work, the storage hosts will open on
its profile the port 9000 (default for minio) to the worker hosts.

This port will be used for now as the standard from the worker
hosts, but if in the future we multiplex storage servers, we will
change that to be configured as sockets (combinations of fqdn and
port numbers).

Change-Id: Ia37946a87497a08f5b86197d749c50deab3de79b