icinga: Performance improvements to icinga-status
This change cuts out >80% of the clock time to run `icinga-status
cumin1001` as measured at alert1001 -- from 2.839s down to 0.502s. That
makes it more feasible to use icinga-status for quick checks (like "is
this hostname known to icinga").

Two changes are involved (and they're fairly inseparable; sorry for the
large diff):

1. We've always had a list of hosts we're interested in, given on the
   command line. Now, we take those hostnames and pass them through to
   _parse_status(). As soon as we find out a block of the status file
   concerns a host that we're interested in, we skip to the next block.
   That way we can skip parsing most of the file, which we were never
   going to read.

2. In order to do that efficiently, we don't loop through the whole file
   line by line -- if we did, we'd still have to scan through those
   blocks we don't care about, in order to look for the close bracket.
   Instead, we start by splitting the whole file into blocks, then parse
   each of *those* line by line. That way, "skip to the next block" is
   fast.

I made a couple of other optimizations, e.g. avoiding most calls to
split() and strip(), and using string slicing instead because it's
faster -- normally I wouldn't bother because it's a little less
readable, but at the center of this hot loop the time savings turned out
to be meaningful.

Bug: T285803
Change-Id: Id924461e6ea3d9a772be6de3862a6c4ee79e573f