mediawiki/maintenance/growthexperiments.pp: Run updateMenteeData every day
The updateMenteeData.php calculates various info about all mentees,
which will be displayed to their mentor via a mentor dashboard.

Because the queries used for the dashboard are analytics-like,
they can easily take several minutes for a large wiki. For that reason,
we've decided to calculate the data daily at a maintenance server,
store them to a database and use the pre-calculated data in the dashboard.

The script should be safe to interupt at any time should that be needed.

Bug: T285811
Change-Id: I6d3773aece4a1800d2cfeffdc15529bc865e5f3e