mediabackup: Update config dir behaviour, add root passwords
Config dir parameter is now deprecated. Keep using it for the
the default TLS cert deployment, but generate the soft links
for the certs on the right location:

/srv/mediabackups/.minio/certs

Delete the old certs subdir.

Also setup a secure user and password, this last one, different
on each datacenter. Setup the regio to be that of the site
(datacenter) where the backup is happening.

Bug: T276442
Change-Id: I23c6923a6b0d120427b0db3abacc48dc286ba24d