cloud galera: have haproxy shut down sessions when marked
First, removing the functionally-unused balance statement.

Since we've seen deadlocks during fail-back because the active server,
came back online, haproxy really should cut them off. There
are potential issues with incomplete transactions as a result of this,
but it is also possible that this will be dealt with correctly by
sqlalchemy, mariadb and/or openstack code (which has a lot of commits
about things that turned out to be bad ideas re:databases).

For concerns I'll just paste the doc and we can think about it. I don't
believe it will be an enormous problem in galera itself, and mysql actually
does handles this by rolling back transactions on disconnect by default:

```
on-marked-up <action>
Modify what occurs when a server is marked up.
Currently one action is available:
- shutdown-backup-sessions: Shutdown sessions on all backup servers. This is
  done only if the server is not in backup state and if it is not disabled
  (it must have an effective weight > 0). This can be used sometimes to force
  an active server to take all the traffic back after recovery when dealing
  with long sessions (e.g. LDAP, SQL, ...). Doing this can cause more trouble
  than it tries to solve (e.g. incomplete transactions), so use this feature
  with extreme care. Sessions killed because a server comes up are logged
  with an 'U' termination code (for "Up").

Actions are disabled by default
```

Bug: T286675
Change-Id: I213a8c35b765a692eff4412abb89c783d7c4724a