elasticsearch: improve check_elasticsearch_shard_size.py
I noticed some icinga UNKNOWNs like the following:

UNKNOWN - int() argument must be a string, a bytes-like object or a number, not 'NoneType'

After debugging a little the check_elasticsearch_shard_size.py script I
noticed that if shards are in INITIALIZING state they report a 'store'
value of None. It seems to me that those shards should not be counted
by the script, but any feedback is welcome.

Change-Id: Ie902a130722f130ba439adc004a95faaec220615