puppetmaster: Stop commits to the private repo with empty messages
Git normally aborts the commit if the commit message is empty. But
because all commits in the private repo are done with sudo, our
commit-msg hook prepends (username), which means the message is *never*
empty. This can surprise users, who intended to abort the commit and
instead end up committing it with a fairly useless message.

With this change, the commit-msg hook exits status 1 in that situation,
aborting the commit.

Change-Id: I8a37c43a0e3aec1e9a7f039f2cb7431abd9641c1