logstash: normalize_level should modify only relevant parts of log obj
All logic in normalize_level:1.0.2 was skipped entirely when a
log object was already defined as a hash to avoid manipulating
it when ECS-compliant.  In cases where the log is structured and
has to be massaged into ECS compliance, the log field could have
entries which are not related to level and trigger the bypass.

Move all instances of normalize_level to ECS pre-filter.

Change-Id: Icd109758c8ff7949380683d81a75a53bb9876e49