metricsinfra: Add config management server
This adds a new role (wmcs::metricinfra::controller), which provisions
the backend uWSGI application that gives the configuration creation
scripts information about configuration (scrape targets, alert rules,
so on) stored in a Trove database.

The deployment model is currently based on hacky git::clone with some
exec's for dependency management, which I dislike as it's tied to Puppet
runs and doesn't give any tools for manual rollouts or rollbacks. I
opened T288052 to fix it in the future.

Add HAProxy configuration for load balancing requests to multiple
backends. Currently the application is fully stateless.

Bug: T286299
Bug: T286300
Change-Id: I9e49e9dad5206767881247c31d4a1b4fae7ebba6