Enable kerberos ticket auto-renewal for a test client
The kstart package includes the binaries `k5start` and `krenew`. We would
like to begin testing the use of krenew in order to reduce the frequency
at which users have to re-enter their Kerberos password with kinit.

This commit adds a new boolean option that can be used to trigger the
automatic renewal of a user's ticket. The enable_autorenew option is
enabled for the analytics_test_cluster::client role and deploys a new
profile fragment. This fragment causes krenew to be called if a ticket
already exists in the user's cache.

Bug: T268985
Change-Id: Id9a6826f83d8ae5b6c705c29c5ae905476713244