gerrit: remove escapeUri
The its-phabricator plugin did not urlencode POST, we worked around that
issue by using `escapeUri`.

I have fixed the plugin upstream to make it urlencode the payload:
https://gerrit-review.googlesource.com/c/plugins/its-phabricator/+/312542

There is thus no need to escape again in the template.

This change should be deployed after the plugin has been upgraded on the
server and before restarting Gerrit.

https://gerrit.wikimedia.org/r/c/operations/software/gerrit/+/705499

Bug: T280197
Change-Id: Id48a04ccf9bc1cb9c1e89345152aa3e4600a4434