dragonfly::dfdaemon
Call dfget with the primary IP to listen on instead of 0.0.0.0 as that
IP is annouced to the supernode as well. Clients receiving chunk
information with 0.0.0.0 will simply try to fetch the chunk from their
local dfget --server instance which in turn will make them fetch the
chunk from the docker-registry directly.
So this is another way to effectively disable P2P while keeping the
system "working".

Bug: T286054
Change-Id: Ib72a66f4240b78e98ae982c4970bdc348d3012e9