maps: decrease tilerator CPU usage in imposm machines
This is a temporarily measure until the cluster is fully migrated.
Reason: Tilerator is generating a large amount of tiles with less
available machines in the cluster, the outcome of this is that tile
generation never end, therefore, CPU usage will be constant.

This patch do the following to mitigate the CPU usage issue:
1) Decrease the available CPU ratio for tilerator from 0.3 to 0.2
2) Trigger tile generatio once a day instead of twice a day

Change-Id: Icf6bf6f08eea8cb1e35d1058606f8530d9572e9c