prometheus: Add hosts_only=false on minio job
We believe the reason why TLS fails to be negotiated is that
the host on the certificate (fqdn) doesn't match the one on the
prometheus config, which is just the hostname.

This would be consistent with the other services that use https.

This should fix scraping on minio https endpoint.

Bug: T276442
Change-Id: Idf75e63c273f7a2bc4970379655359f10281920b