puppetdb: wait for stockpile initialization
On a freshly provisioned host we're attempting to the stockpile queue directory. For this to work
the mount point must exist, but creating it ourselves isn't going to work.

At startup puppetdb doesn't seem to like this fact (it didn't create the 'stockpile' status file and
various other directories) and won't create the files/dirs if they are missing.

Let puppetdb create/init the stockpile instead and only then do the tmpfs mount, with the correct
ownership.

Change-Id: I61bcdcf669846a02913ef2f9bb18e8d4c9490bc1