gerrit: remove SMTP encryption option
3bc1b0906 introduced a Puppet option to set `sendemail.smtpEncryption`.
That comes from a time when we relied on a smart host to send email and
mails actually exited the box.

We never went to enable encryption and nowadays we use `localhost` as a
relay. We have thus no need for setting it.

Also our default 'none', is not the default in Gerrit, it causes
`gerrit.config` to be altered whenever we run `gerrit init`.

Bug: T287122
Change-Id: I43b81298f11445b5efd4a4c9cc62b9011a3c6350